bot/requirements.txt
aiogram
httpx
watchdog
PyJWT